Extract Xiaomi Cluster 0xff01, 0xff02
